refactor(ui): abstract recharts Grid component
