+++ src/components/Visualizations/Graph/Grid.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { CartesianGrid } from "recharts";
+
+const Grid = () => (
+  <CartesianGrid
+    strokeWidth={1}
+    stroke="#2a2a2a"
+    strokeDasharray="3 3"
+    opacity={0.4}
+  />
+);
+
+export default Grid;
